Some refactoring, adding cost functions
