+++ 03-Artificial-Neural-Networks/python/notes
@@ -1,3 +1,4 @@
+Sun Apr 26 11:45:41 BST 2015
 iteration: 180  out of  200
 iteration: 190  out of  200
 Time elapsed  5203.93832898
